fix: map display
