+++ my-japan-next/app/api/google-maps/route.ts
@@ -1,0 +1,7 @@
+import { NextResponse } from 'next/server';
+
+export async function GET() {
+  return NextResponse.json({
+    apiKey: process.env.GOOGLE_MAPS_API_KEY,
+  });
+}
